requirements.txt
pyngrok
pyyaml
requests
subprocess